# back/projects/api.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8081
    description: Generated server url
paths:
...
  /applications:
    post:
      tags:
        - application-controller
      operationId: createApplication
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponseVoid'
...
  /applications/project/{projectId}:
    get:
      tags:
        - application-controller
      operationId: getApplicationsByProject
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: pageable
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponsePageApplicationResponse'
  /applications/my:
    get:
      tags:
        - application-controller
      operationId: getApplicationsByStudent
      parameters:
        - name: pageable
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponsePageApplicationResponse'
components:
  schemas:
...
    BaseResponseVoid:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorResponse'
        success:
          type: boolean
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
    ApplicationRequest:
      required:
        - projectId
      type: object
      properties:
        projectId:
          type: integer
          format: int64
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
...
    PageableObject:
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        unpaged:
          type: boolean
        paged:
          type: boolean
        pageSize:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
...
    SortObject:
      type: object
      properties:
        direction:
          type: string
        nullHandling:
          type: string
        ascending:
          type: boolean
        property:
          type: string
        ignoreCase:
          type: boolean
...
    ApplicationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        memberId:
          type: integer
          format: int64
        projectId:
          type: integer
          format: int64
        status:
          type: string
          enum:
            - PENDING
            - APPROVED
            - REJECTED
        createdAt:
          type: string
          format: date-time
    BaseResponsePageApplicationResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorResponse'
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/PageApplicationResponse'
    PageApplicationResponse:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        pageable:
          $ref: '#/components/schemas/PageableObject'
        numberOfElements:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/ApplicationResponse'
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean